pyproject.toml
[tool.poetry]
name = "rearchive"
version = "0.1.0"
description = ""
authors = ["kureta <skureta@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
tqdm = "^4.66.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
